requirements.txt
absl-py==1.0.0
appnope==0.1.2
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
black==22.1.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
executing==0.8.2
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.14.1
importlib-metadata==4.8.2
interchange==2021.0.4
ipykernel==6.13.0
ipython==8.0.1
ipython-genutils==0.2.0
ipython-sql==0.4.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.2.2
jupyter-core==4.9.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.4.3
matplotlib-inline==0.1.3
monotonic==1.6
mypy-extensions==0.4.3
neo4j==4.4.2
nest-asyncio==1.5.5
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pansi==2020.7.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
prettytable==0.7.2
prompt-toolkit==3.0.27
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py2neo==2021.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pymongo==4.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
sql==0.4.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
stack-data==0.1.4
style==1.1.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==2.0.1
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
update==0.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
